feature(pagination): extract queryFromObject to separate file
