+++ client/utils/queryStringFromObject.js
@@ -1,0 +1,10 @@
+export default (obj) => {
+  let query = '?';
+  if (obj) {
+    const params = Object.keys(obj);
+    params.forEach((param) => {
+      query += `${param}=${obj[param]}&`;
+    });
+  }
+  return query;
+};
